+++ react-app/src/artifacts/ra-view-component.js
@@ -8,6 +8,7 @@ import {AppConstant} from "../app/app-constant";
 import {ApiURL} from "../app/api-url";
 import {RaGsConditionMaker} from "./ra-gs-condition-maker";
 import {RaSelectUtil} from "./ra-select";
+import LoginDialog from "../views/authentication/login-dialog";
 
 
 export default class RaViewComponent extends Component {
@@ -17,6 +18,7 @@ export default class RaViewComponent extends Component {
         this.state = {
             isSystemProgressBarEnabled: false,
             showSystemSnackBar: false,
+            showLoginPopup: false,
             systemSnackBarVariant: "success",
             systemSnackBarMessage: "Empty Message",
             formData: {},
@@ -304,7 +306,8 @@ export default class RaViewComponent extends Component {
             }
         }).catch((error) => {
             if (error.response && error.response.status === 401) {
-                AuthenticationService.logout();
+                this.setState({showLoginPopup: true})
+                // AuthenticationService.logout();
             } else if (failed !== undefined) {
                 failed(error);
             } else {
@@ -378,6 +381,7 @@ export default class RaViewComponent extends Component {
     render() {
         return (
             <React.Fragment>
+                {this.state.showLoginPopup ? (<LoginDialog parent={this}/>) : ""}
                 {RaUtil.showLoader(this.state.isSystemProgressBarEnabled)}
                 <RaSnackBar variant={this.state.systemSnackBarVariant ? this.state.systemSnackBarVariant : "error"}
                             isOpen={this.state.showSystemSnackBar}
